extract init methods into file
